backend/requirements.txt
# AI Sprint Companion - Dependencies

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.0.0

# AI/LLM Integration
openai>=1.0.0

# Document Parsing
python-docx>=0.8.11
lxml>=4.9.0

# HTTP Client (for Jira integration)
httpx>=0.24.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# MCP (Model Context Protocol) for AI Agent integration
mcp>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
ruff>=0.1.0

